Feat: Add basic 404, 400 and 403 error page

+++ orders/views.py
@@ -1,13 +1,20 @@
 from django.shortcuts import render, redirect, reverse, get_object_or_404
 from django.views.generic import ListView, DetailView
 from django.views.generic.edit import CreateView, UpdateView, DeleteView
-from django.contrib.auth.mixins import LoginRequiredMixin
+from django.contrib.auth.mixins import LoginRequiredMixin, UserPassesTestMixin
+from django.contrib.auth.decorators import login_required
 from django.contrib.messages.views import SuccessMessageMixin
 from django.contrib import messages
 from .models import Product
 from .forms import ProductForm
 
 
+class SuperUserRequiredMixin(LoginRequiredMixin, UserPassesTestMixin):
+
+    def test_func(self):
+        return self.request.user.is_superuser
+
+        
 class productsView(ListView):
     model = Product
     template_name = "orders/view_products.html"
@@ -18,7 +25,7 @@ class ProductDetailView(DetailView):
     template_name = "orders/product_detail.html"
 
 
-class AddProduct(SuccessMessageMixin, CreateView):
+class AddProduct(SuperUserRequiredMixin, SuccessMessageMixin, CreateView):
     model = Product
     form_class = ProductForm
     template_name = 'orders/add_product.html'
@@ -33,7 +40,7 @@ class AddProduct(SuccessMessageMixin, CreateView):
         return super().form_valid(form)
 
 
-class EditProduct(LoginRequiredMixin, SuccessMessageMixin, UpdateView):
+class EditProduct(SuperUserRequiredMixin, SuccessMessageMixin, UpdateView):
     model = Product
     form_class = ProductForm
     template_name = 'orders/edit_products.html'
@@ -43,7 +50,13 @@ class EditProduct(LoginRequiredMixin, SuccessMessageMixin, UpdateView):
         return reverse('view_product', kwargs={'slug': self.object.slug})
 
 
+@login_required
 def delete_product(request, slug):
+    # user must be superuser
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, only store owners can do that.')
+        return redirect(reverse('home'))
+
     product = get_object_or_404(Product, slug=slug)
     product.delete()
     messages.success(request, 'Product deleted!')
